perf trace: Handle legacy syscalls
Currently the code skips the first field with the expectation that it is
'nr'. But older kernels do not have the 'nr' field:

    field:int nr;   offset:8;   size:4; signed:1;

Change perf-trace to drop the field if it exists after parsing the format
file.

This fixes the off-by-one problem with older kernels (e.g., RHEL6). e.g,
perf-trace shows this for write:

    3.684 ( 0.007 ms): write(buf: 2, count: 140737077958816  ) = 27

where 2 is really the fd, the huge number is really the buf address, etc.
With this patch you get the more appropriate:

   3.016 ( 0.008 ms): write(fd: 2, buf: 0x7fff9a40f780, count: 26) = 26

Signed-off-by: David Ahern <dsahern@gmail.com>