perf trace: Refactoring to support multiple machine types
perf trace is currently hardcoded to resolve all syscalls based on the
machine type determined by audit_detect_machine (which is based on uname
lookup). However, this fails to properly handle mixed environments --
namely, 32-bit apps on a 64-bit kernel.

Refactor the code to allow multiple machine types to be used. Prep for
the next patch.

Signed-off-by: David Ahern <dsahern@gmail.com>