perf tool: Move callchain option parsing to library code
Code used to parse callchain options is common to perf-record and perf-top
and will be used by perf-trace in an upcoming patch. Since it is common code
it should be in a library / util file rather than the record command.

Code move only; no functional changes. The code in callchain.c is indentical
to what was in builtin-record.c with two exceptions: strings split across
lines was moved to a common line to make checkpatch happy.

Signed-off-by: David Ahern <dsahern@gmail.com>