perf kvm: Use perf_clock for flush time if possible
If perf_clock is available in userspace use it for flush time on the
ordered samples queue.

Get perf_clock timestamp when a round starts. Since perf_clock is
synchronized across all cores within a CPU then no samples can be
generated with a timestamp prior to what is read at the start of the
round. This means that the ordered samples queue can be flushed using that
timestamp without worry.

perf_clock is not guaranteed to be synchronized across sockets. That case
is handled with a heuristic -- subtracting 100 usec from the perf_clock
timesamp read at the beginning of the round and using that as the flush
time.

With this patch 'perf kvm stat live' was able to handle more than 650,000
events per second for several minutes (nested VM) and never failed with
a "Warning: Timestamp below last timeslice flush" error.

Signed-off-by: David Ahern <dsahern@gmail.com>