perf script: add time-of-day option for displaying events
Leverages reference time stored in session header to convert perf_clock
timestamps to time-of-day. Default format is %H:%M:%S and the format can
be changed via the --tod argument to any string recognized by strftime.

This version does *not* use any tracepoints to keep the correlation
between perf_clock and time-of-day up to date. Time discrepancies can
happen if ntpd or a user changes the wallclock.

Signed-off-by: David Ahern <dsahern@gmail.com>